requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.0
pymongo==4.4.1
python-dotenv==1.0.0
numpy==1.24.3
twilio==8.10.0
